AppendListener: add missing is_linked() check
This regression was caused by 8192a30989bdc7aef0621e962bd0a94f8f7768ac
because unlike boost::intrusive::list, IntrusiveList has no implicit
runtime check.

+++ src/AppendListener.hxx
@@ -23,7 +23,8 @@ public:
 	}
 
 	void Unregister() noexcept {
-		siblings.unlink();
+		if (IsRegistered())
+			siblings.unlink();
 	}
 
 	/**
